F&G_index_projet_Barrau_Dalsace_Denis_Walckhoff/gg_trends/requirements.txt
google-cloud-storage
pandas
pytrends
numpy
requests
datetime
robotframework-jsonlibrary